requirements.txt
streamlit
numpy
pillow
opencv-python-headless
ai-edge-litert
plotly
streamlit-lottie
requests